pyproject.toml
[project]
name = "agentic-data-science"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.5.30",
    "ipython>=9.8.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "sqlalchemy-utils>=0.42.1",
    "streamlit>=1.52.2",
]
readme = "README.md"
